requirements.txt
APScheduler==3.5.1
asn1crypto==0.24.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
cx-Oracle==6.4
idna==2.7
lxml==4.2.3
numpy==1.14.5
pandas==0.23.1
paramiko==2.4.1
pexpect==4.6.0
pkg-resources==0.0.0
ptyprocess==0.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
pyfiglet==0.7.5
PyNaCl==1.2.1
PySocks==1.6.8
pyTelegramBotAPI==3.6.3
python-dateutil==2.7.3
python-dotenv==0.8.2
python-ldap==3.1.0
pytz==2018.5
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.9
sshtunnel==0.1.4
tzlocal==1.5.1
urllib3==1.23
